On 64-bit only there's no need to cast to long types

+++ Classes/ActiveCallViewController.m
@@ -169,14 +169,11 @@
     
     if (seconds < 3600) {
         [[self callController] setStatus:[NSString stringWithFormat:@"%02ld:%02ld",
-                                          (unsigned long)((seconds / 60) % 60),
-                                          (unsigned long)(seconds % 60)]];
+                                          (seconds / 60) % 60, seconds % 60]];
     } else {
         [[self callController]
          setStatus:[NSString stringWithFormat:@"%02ld:%02ld:%02ld",
-                    (unsigned long)((seconds / 3600) % 24),
-                    (unsigned long)((seconds / 60) % 60),
-                    (unsigned long)(seconds % 60)]];
+                    (seconds / 3600) % 24, (seconds / 60) % 60, seconds % 60]];
     }
 }
